Added endGame menu, slightly more readable
This commit added an endGame menu for when the game ends. The menu does
not have functionality for the buttons, but it exists.

Also, I made the code a little more readable by somewhat following
JavaDoc rules. And in an attempt to clean up, I surrounded two parts of
the code in about 20 forward-slashes to point them out. I was just
wondering if these are still necessary or have any function in the
final game.